Added parser linkage timeout user option. Parsers that exceed the timeout will now prompt the user for a time extension. Fixed several unhandled NotOnUIThread exceptions. In most cases the processes now switch over to the ui thread if they're not already on it, rather than throwing an exception. Fixed issue where a saved query only began displaying it's dirty state caption after a few minutes. This is a Visual Studio internal anomaly so the document is force saved a second time to kickstart dirty state notifications. Fixed the exception message dialog which had been corrupted during porting. Began a cleanup of redundant classes.

+++ BlackbirdSql.Core/Ctl/ComponentModel/AbstractUomConverter.cs
@@ -235,7 +235,7 @@ public abstract class AbstractUomConverter : TypeConverter, IBEditConverter, IDi
 				if (fmt != null)
 					_UomFmtMax = fmt.FmtRes(_Max);
 
-				_DefaultValue = (int)model[_PropertyName].DefaultValue;
+				_DefaultValue = model[_PropertyName].DefaultValue;
 
 				// The formatted display string of the default value in the current culture.
 				fmt = ResMgr.GetString("ConverterUomFormatDefault_" + _UomKey);
